Update access model for operator|
The operator | can access the matrix either by directly changing
the values (by reference, mm::mutate(M)) or by creating first a
a copy of the matrix (mm::clone(M)).

ToDo:

The order of destruction of the mutate object is not yet well defined,
and therefore it is not yet deterministic enough to work with
expressions like (pseudocode)

  matrix m = a * b * (a | mm::alg::invert)

because operator| (defaults to mutate), should not but could, change
the value of a before the product a * b gets evaluated.